+++ ttbasic/eb_conio.h
@@ -23,8 +23,8 @@ int eb_csize_width(void);
 int eb_pos_x(void);
 int eb_pos_y(void);
 
-unsigned short eb_char_get(int x, int y);
-void eb_char_set(int x, int y, unsigned short c);
+unsigned int eb_char_get(int x, int y);
+void eb_char_set(int x, int y, unsigned int c);
 int eb_cscroll(int x1, int y1, int x2, int y2, int d);
 
 int eb_kbhit(void);
